requirements.txt
# F1 Predictor 2026 - Backend Dependencies
# Python 3.11+

# ============ FastAPI & Web ============
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============ Data Processing ============
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
pyarrow==14.0.1

# ============ API Clients ============
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0

# ============ AI/ML ============
google-generativeai==0.3.0

# ============ Historical Data ============
fastf1==3.2.6

# ============ Development ============
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1
